requirements.txt
langchain-chroma
langchain-core
langchain-openai
pandas
python-dotenv
requests
flask
